[sw-sysemu] [tbox] Fixed BC4 and BC5 formats in TBox emu
Indices at the end of a block were always zero due to the redbits
and greenbits variables being 32 bits, when the should have been
at least 48.

This patch is needed for RTLMIN-1334.


Former-commit-id: 49c7fb5ed946ee4f577d1f2209d5959c8e25b6fe

+++ sw-sysemu/tbox_emu.cpp
@@ -4457,7 +4457,7 @@ void TBOXEmu::decode_BC4(uint8_t *inBuffer, uint8_t *outBuffer, bool signedForma
     float red0, red1;
     float decodedColor[4];
     uint32_t redcode;
-    uint32_t redbits;
+    uint64_t redbits;
 
     //  Convert first reference red color from the compressed block.
     red0 = signedFormat ? fpu::FLT(fpu::sn8_to_f32(inBuffer[0])) : float(inBuffer[0]) * (1.0f / 255.0f);
@@ -4503,7 +4503,7 @@ void TBOXEmu::decode_BC5(uint8_t *inBuffer, uint8_t *outBuffer, bool signedForma
     float red0, red1, green0, green1;
     float decodedColor[4];
     uint32_t redcode, greencode;
-    uint32_t redbits, greenbits;
+    uint64_t redbits, greenbits;
 
     //  Convert first reference red color from the compressed block.
     red0 = signedFormat ? fpu::FLT(fpu::sn8_to_f32(inBuffer[0])) : float(inBuffer[0]) * (1.0f / 255.0f);
